move floatsum construction tests to test_expressions
